(#53) レビュー対応。RedAnimaMovementBehaviourで実装と修正
