+++ Assets/Scenes/TestScene01.unity
@@ -905,7 +905,6 @@ GameObject:
   - component: {fileID: 853969414}
   - component: {fileID: 853969413}
   - component: {fileID: 853969417}
-  - component: {fileID: 853969420}
   m_Layer: 0
   m_Name: Circle
   m_TagString: Untagged
@@ -1057,21 +1056,6 @@ MonoBehaviour:
   CursorTrackingCoefficient: 1
   ThrowStrength: 1
   TrackingSmoothness: 20
---- !u!114 &853969420
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 853969412}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bd1951b41db5524dbc5071d69eb876f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  vectorReverse: 0
-  power: 4
-  speed: 3
 --- !u!1 &893772178
 GameObject:
   m_ObjectHideFlags: 0
